refactor(pytorch): extract prefix trie topology
