+++ lmdeploy/pytorch/paging/block_trie/__init__.py
@@ -1,6 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 """Prefix-cache trie and SSM checkpoint support."""
 
-from .trie import BlockTrie, Node, PrefixCacheStats
+from .node import Node
+from .trie import BlockTrie, PrefixCacheStats
 
 __all__ = ['BlockTrie', 'Node', 'PrefixCacheStats']
