+++ lmdeploy/pytorch/paging/block_trie/checkpoint.py
@@ -24,7 +24,7 @@ import numpy as np
 from lmdeploy.pytorch.messages import PrefixCacheExtraHashes, SchedulerSequence
 
 if TYPE_CHECKING:
-    from .trie import Node
+    from .node import Node
 
 
 StateCheckpointKey: TypeAlias = tuple[str, int, int]
